Added the Shop feature. The player can visit the shop and repair their pick axe for 5 gold.
A Shop object was created and placed on the map in a random location. When the Player is in the shop, a menu is shown which currently only has one option (to repair the pick axe).

Also updated the message to show gold collected and utilized the value instance variable for the Coin object to do this. This is in anticipation of having enemies and treasures drop various amounts of gold.