Began implementing Enemy movement. Each time the player moves, each enemy will move randomly one tile. This was implemented in the "moveEnemy" function in the Enemy class.
The enemy will determine its available moves, and choose one of them randomly. Also updated the image for the Enemy to a pixel art version instead of the placeholder green square.

Next up: "smartly" moving the Enemy towards the player if the player is within a certain number of tiles.