First step to smart Enemy movement implemented. Enemies can now tell whether they can "see" the player or not. Seeing a player means that they are within 3 rows and 3 columns of the player.
Added a row and column attribute to the Tile class. Also added a boolean variable to the Tile class to tell whether the Player is on that Tile.

Added a method to the World class to get the Player tile from the map. moveEnemies now gets the player tile to pass down as parameters to the moveEnemy method in the Enemy class. This information is used to determine whether the Enemy can see the player.

Next up, have the enemy move towards the player if the boolean variable is true, otherwise move randomly.