Didn't think I'd be able to work during 6th period .. BUT ..
Implemented enemy movement so they cannot overlap with other enemies. This turns into a trail of enemies following you. This was done by:

Adding methods to the Enemy class that tell where the player is relative to you if you can see them. If you can, pick one of those directions to move.

Added a boolean variable to Tile to denote whether an enemy is present on that Tile. Only move to a Tile if there is no enemy there.