Initial combat is functional! Player now has 50 HP, 1-2 damage. Enemies have 10HP, 1-2 damage.
Added instance variables to track HP in the Player and Enemy class. Added extra drawString lines to WorldPanel to show the Player and Enemy HP.

One change from the notes above: It will show the Enemy HP if the enemy can see you. Right now it doesn't show which enemy is which.

If the Player gets 0 HP or lower, the game is over. If an Enemy gets 0 HP or lower, it will be removed from the map and will become non-interactable.