Added coins, shop, and key sprites to the "zoomed" in map. Mini-map now shows Enemies as Red, Key and Coin as yellow, and the Shop as Magenta.
Added a "stair tile" for the zoomed in map to show the starting and ending tile.

Big update for today:
The enemy HP bars are now drawn above the enemies on the map in the zoomed in mode. This required making changes to the Entity class so the HP bar can be resized depending on the situation.